+++ Shadows of Autongrad/src/Entities/UI/Toggle.h
@@ -9,8 +9,8 @@ public:
         const uint init_ui_layer = 0);
 
     void Draw() override;
-    inline void MoveBy(Vector2i offset) override { Entity::MoveBy(offset); Move(); }
-    inline void MoveTo(Vector2i new_pos) override { Entity::MoveTo(new_pos); Move(); }
+    inline void MoveBy(Vector2f offset) override { Entity::MoveBy(offset); Move(); }
+    inline void MoveTo(Vector2f new_pos) override { Entity::MoveTo(new_pos); Move(); }
 
 private:
     //Functions
